requirements.txt
construct==2.10.70
impacket==0.12.0
JPype1==1.5.1
pycryptodome==3.19.0
PySide6==6.8.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python_registry==1.3.1
pytz==2023.3.post1
qt_material==2.14
Registry==0.4.2
rich==13.9.4